further work on #4 and #10

+++ include/packet/UDPPacket.h
@@ -5,9 +5,9 @@ class UDPPacket: public IPv4Packet {
     public:
         UDPPacket(char *buf);
         void Extract_UDP_Header(char *buf);
-        int source_port;
-        int dest_port;
-        int length;
-        int checksum;
+        uint16_t source_port;
+        uint16_t dest_port;
+        uint16_t length;
+        uint16_t checksum;
 
 };
